requirements.txt
transformers
torch
scipy
goose3
python-decouple
numpy
flask
hf_xet
gunicorn
